app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.4.0
pymongo==4.6.3
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.0.1
python-multipart==0.0.6
